capgen: produce textgrid for visualization

+++ modules/capgen-1.0/module.py
@@ -42,9 +42,7 @@ def get_length(video_path):
     """
     args = ['ffprobe', '-v', 'error', '-show_entries', 'format=duration',
             '-of', 'default=noprint_wrappers=1:nokey=1', video_path]
-    dur = float(subprocess.check_output(args))
-    LOG.debug('Video length: %s seconds', dur)
-    return dur
+    return float(subprocess.check_output(args))
 
 
 def extract(video_path, temp_dir):
@@ -93,11 +91,12 @@ def extract(video_path, temp_dir):
     return t_arr
 
 
-def get_middle(video_path, t_arr, temp_dir):
+def get_middle(video_path, video_length, t_arr, temp_dir):
     """Get keyframe (middle frame) from each scene.
 
     Arguments:
         video_path: str - path to video file
+        video_length: float - length of video file in seconds
         t_arr: list - list of scene start times from extract()
         temp_dir: str - path to temp directory
     Returns:
@@ -113,14 +112,13 @@ def get_middle(video_path, t_arr, temp_dir):
             frames = json.load(file_)
         LOG.debug('get_middle operation previously completed')
     else:
-        dur = get_length(video_path)
         next_id = 0
-        t_arr.append(dur)
+        t_arr.append(video_length)
         for i in range(1, len(t_arr)):
             t_start = t_arr[i - 1]
             t_middle = t_start + (t_arr[i] - t_start) / 2.
             # fix weird behaviour when a scene is detected near end of a video
-            if dur - t_start < 1.:
+            if video_length - t_start < 1.:
                 break
             next_id += 1
             img_path = os.path.join(temp_dir, '%05d.png' % next_id)
@@ -212,26 +210,77 @@ def predict(frames, temp_dir, neuraltalk_dir):
     return frames
 
 
-def output(frames, file_id, capgen_dir):
+def output(frames, capgen_dir, capgen_file):
     """Produce the output frames and captions.
 
     Arguments:
-        frames: frames data structure from predict()
-        file_id: str - file_id
-        capgen_dir: str - output directory
+        frames: dict - frames data structure from predict()
+        capgen_dir: str - path to output directory
+        capgen_file: str - path to output file
     Returns:
         None
     """
-    capgen_file = os.path.join(capgen_dir, '{}.json'.format(file_id))
-    # copy keyframes to folder and write output file
-    for key, frame in frames.items():
-        path = frame['path']
-        new_path = os.path.join(capgen_dir, key)
-        shutil.copy2(path, new_path)
-        frame['path'] = new_path
-    with open(capgen_file, 'w') as file_out:
-        json.dump(frames, file_out, indent=4, sort_keys=True)
+    # complete check
+    if os.path.exists(capgen_file) and os.path.getsize(capgen_file) > 0:
+        LOG.debug('Previously written %s', capgen_file)
+    else:
+        # copy keyframes to folder and write output file
+        for key in frames:
+            path = frames[key]['path']
+            new_path = os.path.join(capgen_dir, key)
+            shutil.copy2(path, new_path)
+            frames[key]['path'] = new_path
+        with open(capgen_file, 'w') as file_out:
+            json.dump(frames, file_out, indent=4, sort_keys=True)
     LOG.debug('Written %s', capgen_file)
+
+
+def output_to_tg(frames, video_length, capgen_textgrid):
+    """Produce the output TextGrid for visualization.
+
+    Arguments:
+        frames: dict - frames data structure from predict()
+        video_length: float - length of video file
+        capgen_textgrid: str - path to TextGrid file
+    Returns:
+        None
+    """
+    # complete checks
+    if os.path.exists(capgen_textgrid) and os.path.getsize(capgen_textgrid) > 0:
+        LOG.debug('Previously written %s', capgen_textgrid)
+    else:
+        tab4 = ' ' * 4
+        tab8 = ' ' * 8
+        tab12 = ' ' * 12
+        frame = sorted(frames.keys())
+        with open(capgen_textgrid, 'w') as file_out:
+            file_out.write('File type = "ooTextFile"\n')
+            file_out.write('Object class = "TextGrid"\n\n')
+            file_out.write('xmin = 0.0\n')
+            file_out.write('xmax = {}\n'.format(video_length))
+            file_out.write('tiers? <exists>\n')
+            file_out.write('size = 1\n')
+            file_out.write('item []:\n')
+            file_out.write(tab4 + 'item [1]:\n')
+            file_out.write(tab8 + 'class = "IntervalTier"\n')
+            file_out.write(tab8 + 'name = "C"\n')
+            file_out.write(tab8 + 'xmin = 0.0\n')
+            file_out.write(tab8 + 'xmax = {}\n'.format(video_length))
+            file_out.write(tab8 + 'intervals: size = {}\n'.format(len(frames)))
+            frame_cnt = 1
+            while frame_cnt <= len(frames):
+                key = frame[frame_cnt - 1]
+                file_out.write(tab8 + 'intervals [{}]:\n'.format(frame_cnt))
+                file_out.write(
+                    tab12 + 'xmin = {}\n'.format(frames[key]['time']))
+                if frame_cnt < len(frames):
+                    file_out.write(
+                        tab12 + 'xmax = {}\n'.format(frames[frame[frame_cnt]]['time']))
+                else:
+                    file_out.write(tab12 + 'xmax = {}\n'.format(video_length))
+                file_out.write(
+                    tab12 + 'text = "{}"\n'.format(frames[key]['caption']))
+                frame_cnt += 1
 
 
 def capgen(file_id):
@@ -247,16 +296,19 @@ def capgen(file_id):
     if not os.path.exists(capgen_dir):
         os.makedirs(capgen_dir)
     capgen_file = os.path.join(capgen_dir, '{}.json'.format(file_id))
+    capgen_textgrid = os.path.join(capgen_dir, '{}.TextGrid'.format(file_id))
     neuraltalk_dir = os.path.join(CUR_DIR, 'neuraltalk2')
 
     # complete check
-    if os.path.exists(capgen_file):
+    if os.path.exists(capgen_file) and os.path.exists(capgen_textgrid):
         LOG.debug('Previously generated caption for %s', file_id)
     else:
+        video_length = get_length(convert_file)
         t_arr = extract(convert_file, temp_dir)
-        frames = get_middle(convert_file, t_arr, temp_dir)
+        frames = get_middle(convert_file, video_length, t_arr, temp_dir)
         frames = predict(frames, temp_dir, neuraltalk_dir)
-        output(frames, file_id, capgen_dir)
+        output(frames, capgen_dir, capgen_file)
+        output_to_tg(frames, video_length, capgen_textgrid)
 
 
 if __name__ == '__main__':
